Bugfix for image_type and moved instrument_passes_criteria to context.py.
